Add coproc_init_done to avoid races at startup
We were treating coproc_init as "done" once all the capabilities were supplied to the process manager. This was wrong, because the state that ipcd and user programs expect when this is the case requires nsd and coserviced to have completed their startup routines.

Also added code to handle the first usage of the ukern_call_set in a new thread that is not the thread that called coproc_init.