Allow explicitly setting coport otype mapping
libcomsg keeps track of the otypes assigned to different coport types. When coopen returns, it notes the otype of the coport handle and the type of coport opened. This doesn't occur if we get the coport via coexecvec, so we need a function to set it explicitly in those cases.